Rewrite a large part of the library in order to remove `syn`
- The large `macros/lib.rs` was split into multiple smaller files
- To avoid relying on parsing `syn::Expr`, there are declarative macros
  in `bunt` which slightly process the macro input and delegate it to
  the proc macros. In particular, all expressions are wrapped into
  groups, making it easy to skip in the proc macro parser.
- There are not `print[ln]` proc macros anymore as these convenience
  macros can easily be implemented as decl macros in `bunt`.

Despite this huge diff, lots of code stayed the same and was just
moved somewhere.
